Star Ratings & Logos Added

+++ CommentApp-Cores/Models/Comment.cs
@@ -6,7 +6,7 @@ namespace CommentApp_Cores.Models
     {
         public string CommentContent { get; set; }
         public string Title { get; set; }
-        public float Point { get; set; }
+        public int Point { get; set; }
         public Category Category { get; set; }
         public int CategoryId { get; set; }
     }
